feat: improve v2 migration support
(cherry picked from commit 3c87e775ed02ff099abe8c9865c91f57de56b592)

docs: fix v2board migration command

(cherry picked from commit 571c12b79ef03bb3431c7c62ef36924ae90b804a)

+++ database/migrations/2025_01_10_152139_add_device_limit_column.php
@@ -8,24 +8,40 @@ return new class extends Migration {
     public function up(): void
     {
         Schema::table('v2_plan', function (Blueprint $table) {
-            $table->unsignedInteger('device_limit')->nullable()->after('speed_limit');
+            if (!Schema::hasColumn('v2_plan', 'device_limit')) {
+                $table->unsignedInteger('device_limit')->nullable()->after('speed_limit');
+            }
         });
         Schema::table('v2_user', function (Blueprint $table) {
-            $table->integer('device_limit')->nullable()->after('expired_at');
-            $table->integer('online_count')->nullable()->after('device_limit');
-            $table->timestamp('last_online_at')->nullable()->after('online_count');
+            if (!Schema::hasColumn('v2_user', 'device_limit')) {
+                $table->integer('device_limit')->nullable()->after('expired_at');
+            }
+            if (!Schema::hasColumn('v2_user', 'online_count')) {
+                $table->integer('online_count')->nullable()->after('device_limit');
+            }
+            if (!Schema::hasColumn('v2_user', 'last_online_at')) {
+                $table->timestamp('last_online_at')->nullable()->after('online_count');
+            }
         });
     }
 
     public function down(): void
     {
         Schema::table('v2_user', function (Blueprint $table) {
-            $table->dropColumn('device_limit');
-            $table->dropColumn('online_count');
-            $table->dropColumn('last_online_at');
+            if (Schema::hasColumn('v2_user', 'device_limit')) {
+                $table->dropColumn('device_limit');
+            }
+            if (Schema::hasColumn('v2_user', 'online_count')) {
+                $table->dropColumn('online_count');
+            }
+            if (Schema::hasColumn('v2_user', 'last_online_at')) {
+                $table->dropColumn('last_online_at');
+            }
         });
         Schema::table('v2_plan', function (Blueprint $table) {
-            $table->dropColumn('device_limit');
+            if (Schema::hasColumn('v2_plan', 'device_limit')) {
+                $table->dropColumn('device_limit');
+            }
         });
     }
 };
